+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 # draw-graph Changelog
 
-## [Unreleased]
-### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+## [1.0.0]
+
+- 1.01 Maven dependencies graph with mermaid.js
+- 1.00 Method call graph with mermaid.js
+
+# 中文更新日志
+
+- 1.01 Maven 关系图 mermaid.js 版
+- 1.00 方法调用关系图 mermaid.js 版